Moved NetworkObject types into models folder, other code review tweaks
